+++ packages/vue-tab-router/src/components/keep-alive-enhance/core/apiAsyncComponent.ts
@@ -1,4 +1,0 @@
-import { ComponentInternalInstance, VNode, ComponentOptions } from "vue";
-
-export const isAsyncWrapper = (i: ComponentInternalInstance | VNode): boolean =>
-  !!(i.type as ComponentOptions).__asyncLoader;