+++ packages/vue-tab-router/src/components/keep-alive-enhance/core/componentPublicInstance.ts
@@ -1,6 +1,0 @@
-import { ComponentInternalInstance } from "vue";
-
-export interface ComponentRenderContext {
-  [key: string]: any;
-  _: ComponentInternalInstance;
-}