+++ packages/vue-tab-router/src/tab.ts
@@ -1,4 +1,4 @@
-import { TabGuard, TabGuardName } from "./types";
+import { PersistedTab, TabCloseGuard, TabEnterGuard, TabLeaveGuard } from "./types";
 
 export class Tab {
   _id!: string;
@@ -10,17 +10,32 @@ export class Tab {
   _single?: boolean;
   _noCache?: boolean;
   _loading?: boolean;
-  // _onBeforeTabOpen?: TabGuard;
-  _onBeforeTabEnter?: TabGuard;
-  _onBeforeTabLeave?: TabGuard;
-  _onBeforeTabClose?: TabGuard;
+  _onBeforeTabEnter?: TabEnterGuard;
+  _onBeforeTabLeave?: TabLeaveGuard;
+  _onBeforeTabClose?: TabCloseGuard;
 
   viewName?: string;
   viewIcon?: string;
   viewUrl!: string;
   viewProps?: Record<string, any>;
 
-  constructor(options?: Omit<Tab, TabGuardName>) {
+  constructor(options?: Partial<PersistedTab>) {
     Object.assign(this, options || {});
+  }
+
+  toJSON(): PersistedTab {
+    return {
+      _id: this._id,
+      _sourceId: this._sourceId,
+      _isActive: this._isActive,
+      _noClose: this._noClose,
+      _isFirst: this._isFirst,
+      _single: this._single,
+      _noCache: this._noCache,
+      viewName: this.viewName,
+      viewIcon: this.viewIcon,
+      viewUrl: this.viewUrl,
+      viewProps: this.viewProps,
+    };
   }
 }